Expose StreamConfiguration
