--- client/coral-embed-stream/src/tabs/configure/components/Checkbox.css
@@ -1,59 +1,0 @@
-.root {
-  position: relative;
-  display: inline-block;
-}
-
-.input {
-  position: absolute;
-  left: 7px;
-  bottom: 7px;
-  margin: 0;
-  padding: 0;
-  outline: none;
-  cursor: pointer;
-  pointer-events: none;
-  opacity: 0;
-}
-
-.checkbox {
-  cursor: pointer;
-}
-
-.checkbox:before {
-  content: "\e835";
-  color: #717171;
-  left: 4px;
-  top: 0px;
-  width: 18px;
-  height: 18px;
-  font-family: 'Material Icons';
-  font-weight: normal;
-  font-style: normal;
-  font-size: 24px;
-  line-height: 1;
-  text-transform: none;
-  letter-spacing: normal;
-  word-wrap: normal;
-  white-space: nowrap;
-  direction: ltr;
-  vertical-align: -6px;
-  text-rendering: optimizeLegibility;
-  font-feature-settings: 'liga';
-  transition: all .2s ease;
-  z-index: 1;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-.checkboxChecked:before {
-  content: "\e834";
-  color: #00a291;
-}
-
-.input:focus + .checkbox:before {
-  color: #00a291;
-}
-
-.input:focus + .checkboxChecked:before {
-  color: #00e291;
-}
