+++ client/coral-framework/components/StreamConfiguration.css
