+++ plugin-api/beta/client/components/index.js
@@ -7,3 +7,4 @@ export {default as CommentTimestamp} from 'coral-framework/components/CommentTim
 export {default as CommentDetail} from 'coral-framework/components/CommentDetail';
 export {default as CommentContent} from 'coral-framework/components/CommentContent';
 export {default as ConfigureCard} from 'coral-framework/components/ConfigureCard';
+export {default as StreamConfiguration} from 'coral-framework/components/StreamConfiguration';
